LPD-47877 Move util class to api module
